+++ Extensions/AudioScripts/Mpdn.DynamicRangeCompressor.cs
@@ -29,35 +29,32 @@ namespace Mpdn.Extensions.AudioScripts
 {
     namespace Mpdn
     {
-        public class DynamicRangeCompressor : AudioChain
-        {
-            public float ThresholddB { get; set; }
-            public float Ratio { get; set; }
-            public float MakeupGaindB { get; set; }
-            public int AttackMs { get; set; }
-            public int ReleaseMs { get; set; }
-
-            public DynamicRangeCompressor()
-            {
-                ThresholddB = -15;
-                Ratio = 3;
-                MakeupGaindB = 3;
-                AttackMs = 200; // 0.2s
-                ReleaseMs = 1000; // 1s
-            }
-
+        public class DynamicRangeCompressorFilter : AudioFilter { 
             // DC offset to prevent denormal
             protected const float DC_OFFSET = 1.0e-25f;
 
             private float m_EnvdB = DC_OFFSET;
-            private EnvelopeDetector m_Attack;
-            private EnvelopeDetector m_Release;
-            private int m_SampleRate;
-            private int m_AttackMs;
-            private int m_ReleaseMs;
-            private int m_SampleCount;
 
             private float[] m_DevOverdBs;
+
+            private readonly float ThresholddB;
+            private readonly float Ratio;
+            private readonly float MakeupGaindB;
+            private readonly int m_AttackMs;
+            private readonly int m_ReleaseMs;
+
+            private EnvelopeDetector m_Attack;
+            private EnvelopeDetector m_Release;
+            private int m_SampleCount;
+
+            public DynamicRangeCompressorFilter(float thresholddB, float ratio, float makeupGaindB, int attackMs, int releaseMs)
+            {
+                m_AttackMs = attackMs;
+                m_ReleaseMs = releaseMs;
+                ThresholddB = thresholddB;
+                Ratio = ratio;
+                MakeupGaindB = makeupGaindB;
+            }
 
             protected override void OnLoadAudioKernel()
             {
@@ -72,18 +69,17 @@ namespace Mpdn.Extensions.AudioScripts
 
             protected override void Process(float[,] samples, short channels, int sampleCount)
             {
-                var sampleRate = Input.Format.nSamplesPerSec;
-                if (m_SampleRate != sampleRate || m_AttackMs != AttackMs || m_ReleaseMs != ReleaseMs)
-                {
-                    m_Attack = new EnvelopeDetector(AttackMs, sampleRate);
-                    m_Release = new EnvelopeDetector(ReleaseMs, sampleRate);
-                    m_SampleRate = sampleRate;
-                    m_AttackMs = AttackMs;
-                    m_ReleaseMs = ReleaseMs;
-                    m_EnvdB = DC_OFFSET;
-                }
+                Initialize();
 
                 Compress(samples, sampleCount, ThresholddB, Ratio, MakeupGaindB);
+            }
+
+            protected override void Initialize()
+            {
+                var sampleRate = Input.Format.nSamplesPerSec;
+                m_Attack = new EnvelopeDetector(m_AttackMs, sampleRate);
+                m_Release = new EnvelopeDetector(m_ReleaseMs, sampleRate);
+                m_EnvdB = DC_OFFSET;
             }
 
             private void Compress(float[,] samples, int sampleCount, float thresholddB, float ratio, float makeupGaindB)
@@ -139,6 +135,29 @@ namespace Mpdn.Extensions.AudioScripts
                 DisposeGpuResources();
                 m_DevOverdBs = Gpu.Allocate<float>(sampleCount);
                 m_SampleCount = sampleCount;
+            }
+        }
+
+        public class DynamicRangeCompressor : AudioChain<DynamicRangeCompressorFilter>
+        {
+            public float ThresholddB { get; set; }
+            public float Ratio { get; set; }
+            public float MakeupGaindB { get; set; }
+            public int AttackMs { get; set; }
+            public int ReleaseMs { get; set; }
+
+            public DynamicRangeCompressor()
+            {
+                ThresholddB = -15;
+                Ratio = 3;
+                MakeupGaindB = 3;
+                AttackMs = 200; // 0.2s
+                ReleaseMs = 1000; // 1s
+            }
+
+            protected override DynamicRangeCompressorFilter MakeFilter()
+            {
+                return new DynamicRangeCompressorFilter(ThresholddB, Ratio, MakeupGaindB, AttackMs, ReleaseMs);
             }
         }
